+++ include/Callable.hpp
@@ -10,7 +10,7 @@ class Interpreter;
 class Callable
 {
 public:
-    virtual int getArity() const = 0;
+    virtual size_t getArity() const = 0;
     virtual std::any call(Interpreter& interpreter, const std::vector<std::any>& args) const = 0;
     virtual std::string toString() const = 0;
     virtual ~Callable() = default;
